+++ packages/amplify-cli-auth/src/commands/logout.js
@@ -15,7 +15,7 @@ export default {
 	},
 	async action({ argv, cli, console }) {
 		const [
-			{ initSDK },
+			{ initSDK, isHeadless },
 			{ default: snooplogg }
 		] = await Promise.all([
 			import('@axway/amplify-cli-utils'),
@@ -26,13 +26,23 @@ export default {
 			argv.all = true;
 		}
 
+		const { highlight, warning } = snooplogg.styles;
+
 		const { sdk } = initSDK({
 			baseUrl:  argv.baseUrl,
 			clientId: argv.clientId,
 			env:      argv.env,
 			realm:    argv.realm
 		});
-		const revoked = await sdk.auth.logout(argv);
+		const revoked = await sdk.auth.logout({
+			...argv,
+			onOpenBrowser() {
+				if (isHeadless()) {
+					console.log(warning(' ┃ Logging out of a platform account requires a web browser and is unsupported'));
+					console.log(warning(' ┃ in headless environments.\n'));
+				}
+			}
+		});
 
 		await cli.emitAction('axway:auth:logout', revoked);
 
@@ -43,15 +53,14 @@ export default {
 
 		// pretty output
 		if (revoked.length) {
-			const { highlight } = snooplogg.styles;
 			console.log('Revoked authenticated accounts:');
 			for (const account of revoked) {
 				console.log(` ${highlight(account.name)}`);
 			}
 		} else if (Array.isArray(argv.accounts) && argv.accounts.length === 1) {
-			console.log(`No account "${argv.accounts[0]}" to revoke.`);
+			throw new Error(`No account "${argv.accounts[0]}" found`);
 		} else {
-			console.log('No accounts to revoke.');
+			throw new Error('No authenticatd accounts found');
 		}
 	}
 };